requirements.txt
python-telegram-bot>=20.0
apscheduler>=3.10
aiosqlite>=0.19
